Refined two Test Scenarios working with the user interface using special libraries. The first Test Scenario creates a new contact with correct data, and the second Test Scenario creates a new contact with blank data.
The saved Test Scenarios were brought to a more readable appearance. Transformed by refactoring means, extracted auxiliary methods, made these methods parameterized and grouped them into a special class called Contact in the contact.py file. Now instead of sending a large number of individual parameters (and there are as many as 24!), we shape an object and send this object as a parameter.